refactor: Create different generator modules
