+++ src/world/generation/generators.rs
@@ -1,0 +1,2 @@
+pub mod default;
+pub mod fixed;
